restructured cypress folder structure
- it now follows the suggested structure more closely
- tests are not in an 'integration' folder anymore, as they were not
integration tests, but rather e2e
- helper functions were moved to the support folder
- they are not defined on cypress itself as suggested in the example
'commands.ts' file, but rather imported directly in test files
