+++ project-base/storefront/cypress/docker/Dockerfile
@@ -6,4 +6,4 @@ COPY --chown=node:node . /app
 
 USER node
 
-CMD npm ci && npx cypress run
+CMD npm ci && npx cypress run
