+++ project-base/storefront/cypress/e2e/cart/productAddToCartFromBrandList.cy.ts
@@ -1,7 +1,7 @@
 import { brandSencor, products, totalPrice, url } from 'fixtures/demodata';
-import { checkProductInCart, checkTotalPriceInCart } from 'integration/Functions/cart';
-import { checkProductAndGoToCartFromCartPopupWindow } from 'integration/Functions/cartPopupWindow';
-import { addProductToCartFromProductList } from 'integration/Functions/productList';
+import { checkProductInCart, checkTotalPriceInCart } from 'support/cart';
+import { checkProductAndGoToCartFromCartPopupWindow } from 'support/cartPopupWindow';
+import { addProductToCartFromProductList } from 'support/productList';
 
 it('Brand list - Adding product to cart from brand list and check product in cart', () => {
     cy.visit(url.brandOverwiev);
