+++ project-base/storefront/cypress/e2e/cart/productAddToCartFromProductVariant.cy.ts
@@ -1,7 +1,7 @@
 import { products, totalPrice, url } from 'fixtures/demodata';
-import { checkProductInCart, checkTotalPriceInCart } from 'integration/Functions/cart';
-import { checkProductAndGoToCartFromCartPopupWindow } from 'integration/Functions/cartPopupWindow';
-import { addProductVariantToCartFromProductDetail } from 'integration/Functions/productDetail';
+import { checkProductInCart, checkTotalPriceInCart } from 'support/cart';
+import { checkProductAndGoToCartFromCartPopupWindow } from 'support/cartPopupWindow';
+import { addProductVariantToCartFromProductDetail } from 'support/productDetail';
 
 it('Product variant - Adding variant product to cart from product detail and check product in cart', () => {
     cy.visit(products.philips32PFL4308.url);
