+++ project-base/storefront/cypress/support/cartPopupWindow.ts
