+++ project-base/storefront/cypress/support/cookies.ts
