+++ project-base/storefront/cypress/support/header.ts
