+++ project-base/storefront/cypress/support/productDetail.ts
